Move CalculateMaterialReward
